Fixing group erros
